Simplify: code-based routing, remove dead deps, merge endpoints
Switch from file-based to code-based TanStack Router routing:
- Delete routes/ directory + auto-generated routeTree.gen.ts
- Remove @tanstack/router-plugin devDependency + TanStackRouterVite plugin
- Two source files: main.tsx (routes + root layout) + artifact.tsx (view)
- Route definitions inline in main.tsx using createRoute/createRootRoute

Other simplifications:
- Remove unused @codemirror/merge dependency + dead merge CSS from index.html
- Merge /api/file + /api/blob into single /api/blob endpoint (defaults to HEAD)
- Use Response.json() instead of custom json() helper
- Drop useCallback wrappers (no memoized children to benefit)
- Inline one-use style constants, keep only H3 shared
- Primary export at top of worker, helpers below (per RULES.md)
