+++ apps/artifacts/package.json
@@ -17,13 +17,11 @@
     "@codemirror/language": "^6.10.0",
     "@codemirror/view": "^6.35.0",
     "@codemirror/language-data": "^6.5.1",
-    "@codemirror/merge": "^6.12.1",
     "isomorphic-git": "^1.27.0"
   },
   "devDependencies": {
     "@cloudflare/vite-plugin": "^1.29.1",
     "@cloudflare/workers-types": "^4.20260317.1",
-    "@tanstack/router-plugin": "^1.120.0",
     "@vitejs/plugin-react": "^4.4.1",
     "vite": "^6.3.5",
     "typescript": "^5.8.3",
