--- apps/artifacts/src/routeTree.gen.ts
@@ -1,77 +1,0 @@
-/* eslint-disable */
-
-// @ts-nocheck
-
-// noinspection JSUnusedGlobalSymbols
-
-// This file was automatically generated by TanStack Router.
-// You should NOT make any changes in this file as it will be overwritten.
-// Additionally, you should also exclude this file from your linter and/or formatter to prevent it from being checked or modified.
-
-import { Route as rootRouteImport } from './routes/__root'
-import { Route as ArtifactRouteImport } from './routes/$artifact'
-import { Route as IndexRouteImport } from './routes/index'
-
-const ArtifactRoute = ArtifactRouteImport.update({
-  id: '/$artifact',
-  path: '/$artifact',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const IndexRoute = IndexRouteImport.update({
-  id: '/',
-  path: '/',
-  getParentRoute: () => rootRouteImport,
-} as any)
-
-export interface FileRoutesByFullPath {
-  '/': typeof IndexRoute
-  '/$artifact': typeof ArtifactRoute
-}
-export interface FileRoutesByTo {
-  '/': typeof IndexRoute
-  '/$artifact': typeof ArtifactRoute
-}
-export interface FileRoutesById {
-  __root__: typeof rootRouteImport
-  '/': typeof IndexRoute
-  '/$artifact': typeof ArtifactRoute
-}
-export interface FileRouteTypes {
-  fileRoutesByFullPath: FileRoutesByFullPath
-  fullPaths: '/' | '/$artifact'
-  fileRoutesByTo: FileRoutesByTo
-  to: '/' | '/$artifact'
-  id: '__root__' | '/' | '/$artifact'
-  fileRoutesById: FileRoutesById
-}
-export interface RootRouteChildren {
-  IndexRoute: typeof IndexRoute
-  ArtifactRoute: typeof ArtifactRoute
-}
-
-declare module '@tanstack/react-router' {
-  interface FileRoutesByPath {
-    '/$artifact': {
-      id: '/$artifact'
-      path: '/$artifact'
-      fullPath: '/$artifact'
-      preLoaderRoute: typeof ArtifactRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/': {
-      id: '/'
-      path: '/'
-      fullPath: '/'
-      preLoaderRoute: typeof IndexRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-  }
-}
-
-const rootRouteChildren: RootRouteChildren = {
-  IndexRoute: IndexRoute,
-  ArtifactRoute: ArtifactRoute,
-}
-export const routeTree = rootRouteImport
-  ._addFileChildren(rootRouteChildren)
-  ._addFileTypes<FileRouteTypes>()
