--- apps/artifacts/src/routes/index.tsx
@@ -1,17 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/")({
-  component: () => (
-    <div
-      style={{
-        flex: 1,
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center",
-        color: "#8b949e",
-      }}
-    >
-      Select a repo to get started
-    </div>
-  ),
-});
