+++ apps/artifacts/vite.config.ts
@@ -1,9 +1,7 @@
-// https://tanstack.com/router/latest/docs/framework/react/start/getting-started#vite-configuration
-import { TanStackRouterVite } from "@tanstack/router-plugin/vite";
 import react from "@vitejs/plugin-react";
 import { cloudflare } from "@cloudflare/vite-plugin";
 import { defineConfig } from "vite";
 
 export default defineConfig({
-  plugins: [TanStackRouterVite(), react(), cloudflare()],
+  plugins: [react(), cloudflare()],
 });
